Refactorized partial deviations.
Piano chords are clean now, the are not levigated any more due to the beat
effect that appeared because sound generators that needed to be lengthened
to accomodate for an interpolated generator with more partials, but got
even-distance partials disregarding string inharmonicity.

For this to work, SPREAD in instrument definitions specifies the
cumulative deviations of partial frequences. This SPREAD is inherited by
both parent sound generators in an interpolation based on weighted
average.

Furthermore, PARTIALS property in instrument definition has been renamed
to "PROFILE", because "PARTIALS" would suit as well to "TIMBRE" and
"SPREAD", so the name turns out to be badly chosen.