Delayed stress bound shape calculation and rendering.
Plus, fixed a prev_run_cache-related bug related
to length_secs that we have to store manually in
the note object on retrieval.

Finally, some more work on piano sound.
Currently not much hope that it ever gets close to
a real piano. It still resembles a cheap keyboard.
But I will regain my confidence and continue designing.